Move helpers to integrations/firebase
